some W4 level warnings fixed; some x64 warnings fixed;

+++ src/engine/core/path.h
@@ -76,7 +76,7 @@ public:
 	uint32 getHash() const { return m_data->m_id; }
 
 	const char* c_str() const { return m_data->m_path; }
-	size_t length() const { return stringLength(m_data->m_path); }
+	int length() const { return stringLength(m_data->m_path); }
 	bool isValid() const { return m_data->m_path[0] != '\0'; }
 
 private:
